requirements.txt
c2pa-python>=0.28.0
cryptography
requests
typer 
rich 
pillow